+++ Classes/Private/Manager/Database/OCTRealmManager.m
@@ -275,30 +275,49 @@ static NSString *kSettingsStorageObjectPrimaryKey = @"kSettingsStorageObjectPrim
     return call;
 }
 
-- (void)removeChatWithAllMessages:(OCTChat *)chat
+- (void)removeMessages:(NSArray<OCTMessageAbstract *> *)messages
+{
+    NSParameterAssert(messages);
+
+    OCTLogInfo(@"removing messages %lu", (unsigned long)messages.count);
+
+    dispatch_sync(self.queue, ^{
+        [self.realm beginWriteTransaction];
+
+        NSMutableSet *changedChats = [NSMutableSet new];
+        for (OCTMessageAbstract *message in messages) {
+            [changedChats addObject:message.chatUniqueIdentifier];
+        }
+
+        [self removeMessagesWithSubmessages:messages];
+
+        for (NSString *chatUniqueIdentifier in changedChats) {
+            RLMResults *messages = [OCTMessageAbstract objectsInRealm:self.realm where:@"chatUniqueIdentifier == %@", chatUniqueIdentifier];
+            messages = [messages sortedResultsUsingProperty:@"dateInterval" ascending:YES];
+
+            OCTChat *chat = [OCTChat objectInRealm:self.realm forPrimaryKey:chatUniqueIdentifier];
+            chat.lastMessage = messages.lastObject;
+        }
+
+        [self.realm commitWriteTransaction];
+    });
+}
+
+- (void)removeAllMessagesInChat:(OCTChat *)chat removeChat:(BOOL)removeChat
 {
     NSParameterAssert(chat);
 
     OCTLogInfo(@"removing chat with all messages %@", chat);
 
     dispatch_sync(self.queue, ^{
-        RLMResults *messages = [OCTMessageAbstract objectsInRealm:self.realm where:@"chat == %@", chat];
+        RLMResults *messages = [OCTMessageAbstract objectsInRealm:self.realm where:@"chatUniqueIdentifier == %@", chat.uniqueIdentifier];
 
         [self.realm beginWriteTransaction];
-        for (OCTMessageAbstract *message in messages) {
-            if (message.messageText) {
-                [self.realm deleteObject:message.messageText];
-            }
-            if (message.messageFile) {
-                [self.realm deleteObject:message.messageFile];
-            }
-            if (message.messageCall) {
-                [self.realm deleteObject:message.messageCall];
-            }
-        }
 
-        [self.realm deleteObjects:messages];
-        [self.realm deleteObject:chat];
+        [self removeMessagesWithSubmessages:messages];
+        if (removeChat) {
+            [self.realm deleteObject:chat];
+        }
 
         [self.realm commitWriteTransaction];
     });
@@ -476,6 +495,24 @@ static NSString *kSettingsStorageObjectPrimaryKey = @"kSettingsStorageObjectPrim
     }];
 
     return messageAbstract;
+}
+
+// Delete an NSArray, RLMArray, or RLMResults of messages from this Realm.
+- (void)removeMessagesWithSubmessages:(id)messages
+{
+    for (OCTMessageAbstract *message in messages) {
+        if (message.messageText) {
+            [self.realm deleteObject:message.messageText];
+        }
+        if (message.messageFile) {
+            [self.realm deleteObject:message.messageFile];
+        }
+        if (message.messageCall) {
+            [self.realm deleteObject:message.messageCall];
+        }
+    }
+
+    [self.realm deleteObjects:messages];
 }
 
 @end